+++ iiif_now/__init__.py
@@ -2,3 +2,4 @@ from .csvreader import DataReader
 from .datacanvas import DataCanvas
 from .manifest import ANManifest
 from .navplace import NavPlace
+from .thumbnail import Thumbnail